factor out name and code validation
