+++ src/settings/validateNameAndCode.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import { FormattedMessage } from 'react-intl';
+
+export default function validateNameAndCode(item) {
+  const errors = {};
+  if (!item.code) {
+    errors.code = <FormattedMessage id="ui-inventory.fillIn" />;
+  }
+  if (!item.name) {
+    errors.name = <FormattedMessage id="ui-inventory.fillIn" />;
+  }
+  return errors;
+}
